+++ src/PlayGifSettingsViewModel.cs
@@ -101,6 +101,16 @@ namespace PlayGif
             _plugin.RunLayoutReport();
         }
 
+        public void ConvertAllToMp4()
+        {
+            _plugin.RunBulkMp4Conversion();
+        }
+
+        public void RepairDescriptionLinks()
+        {
+            _plugin.RunDescriptionRepair();
+        }
+
         public void OpenLogFolder()
         {
             _plugin.OpenLogFolderPublic();